feat(DynEntity): land upstream client/load_obj/pieces into POSIX/Switch build
Wire src/DynEntity/DynEntity_client.cpp, DynEntity_load_obj.cpp, and
DynEntity_pieces.cpp into both backends — the per-client DynEntity
state, the BSP-driven entity loader (with save/restore), and the
gib/piece spawn pipeline.

Source fixes for -Werror clean build:
  - Fix non-portable include path: `dynentity_client.h` → the actual
    file name (case-sensitive on POSIX filesystems).
  - Add explicit parens around `&` inside `|` in the rotated-bounds
    select-and-combine expressions (modelBoundsFloat4 and the
    bounds/HIDWORD scratch in DynEntCl_TraceAllBounds).
  - KISAKHACK uintptr_t bridges everywhere the hex-rays decompiler
    truncated dxBody pointers through int32_t physObjId fields
    (DynEntity_client / DynEntity_pieces / DynEntity_load_obj).
  - [[maybe_unused]] on the set-but-not-used dynEntClient scratch in
    DynEntCl_DrawCollision.

Compat shim:
  - Add sscanf_s alias to sscanf in kisak_compat.h (DynEntity_load_obj
    parses the loaded entity definitions with sscanf_s).

Dup removal in posix_backbone_stubs:
  - Comment out 18 DynEnt_* / DynEntCl_* stubs now provided by upstream.

Satellite stubs (so the DynEntity batch links until renderer/physics
subsystems land):
  - R_LinkDynEnt / R_UnlinkDynEnt.
  - Phys_AddJitterRegion / Phys_ObjSetAngularVelocity.
  - XModelPiecesPrecache (returns nullptr).
  - R_FilterXModelIntoScene.

+++ src/DynEntity/DynEntity_pieces.cpp
@@ -59,7 +59,7 @@ void __cdecl DynEntPieces_AddDrawSurfs()
             Sys_EnterCriticalSection(CRITSECT_PHYSICS);
             Phys_ObjGetInterpolatedState(
                 PHYS_WORLD_FX,
-                (dxBody *)g_breakablePieces[i].physObjId,
+                (dxBody *)(uintptr_t)g_breakablePieces[i].physObjId,
                 placement.base.origin,
                 placement.base.quat);
             placement.scale = 1.0;
@@ -132,7 +132,7 @@ bool __cdecl DynEntPieces_SpawnPhysicsModel(
         angularVelocity[2] = dynEntPieces_angularVelocity->current.vector[2];
         velocity[2] = velocity[2] + model->physPreset->piecesUpwardVelocity;
         Sys_EnterCriticalSection(CRITSECT_PHYSICS);
-        physObjId = (int)DynEntPieces_SpawnPhysObj(
+        physObjId = (int)(uintptr_t)DynEntPieces_SpawnPhysObj(
             model->name,
             mins,
             maxs,
@@ -147,7 +147,7 @@ bool __cdecl DynEntPieces_SpawnPhysicsModel(
             DynEntPieces_CalcForceDir(hitDir, model->physPreset->piecesSpreadFraction, forceDir);
             Phys_ObjBulletImpact(
                 PHYS_WORLD_DYNENT,
-                (dxBody *)physObjId,
+                (dxBody *)(uintptr_t)physObjId,
                 hitPos,
                 forceDir,
                 dynEntPieces_impactForce->current.value,
